vgrep: remove internal SLOC class
The SLOC class has been used to represent a source line of code that
matched the specified pattern.  This class also entailed a considerable
memory overhead that can easily lead to freeze the system when millions
of lines match the specified pattern.

This change removes the class entirely.  The file, line, and content of
the line are now saved as simple triples (file, line, content).