execute(): return output when `failok` is true
Replace `check_output()` with `Popen()` and return the output of the
failed command when `failok` is set. This didn't work before as
`check_output()` raises an error instead.

I am not sure why and how this issue didn't show up earlier. Apparently
I had to vgrep kubernetes first, which isn't fun with grep at all, which
returns with 2 on my Leap 42.2 machine when running in kubernertes.

Signed-off-by: Valentin Rothberg <vrothberg@suse.com>