Avoid race conditions when reading/writing cache file
For large cache files and slow file systems (like NFS) vgrep
sometimes shows an error on a subsequent 'show' command:

$> vgrep "something"
$> vgrep -s 1337
Error loading cache: unexpected end of JSON input

Fix this by waiting for the cacheWrite() goroutine to finish
before exiting the first vgrep call and explicitely syncing
the cache file.

As this error could also happen if multiple instances of
vgrep are run in parallel (and write to the cache file
concurrently), this additionally requires a lockfile which
is locked when writing to or reading from the cache.

Signed-off-by: Andreas Ziegler <andreas.ziegler@fau.de>