Enable basic functionality on windows
Lock directory and cache directory creation were failing when compiling
and running on windows. This fixes the paths to be the windows
equivalents when running on windows.

After this change basic functionality works, on my machine I don't get
any color unless I use the /no-less option, but that's fine.