vgrep: ignore non-matching lines
As reported in #84 `vgrep` panic'd when running with `-{A,B,C}` options
for grep.  These options print context lines and yield a different
format for the separator lines, which caused panics in the rather
optimistic parser.

Address the panics by ignoring lines that do not match the format. Those
lines will be skipped and can be found in the `--debug` log.

Fixes: #84
Reported-by: @marcosnils
Signed-off-by: Valentin Rothberg <rothberg@redhat.com>