README: clarify search/show
Clarify that `vgrep` doesn't allow for searching and showing in one
command.  The technical reason for that is to allow for a more
user-friendly way of using the `--show` command.  In particular,
to support `--show x y z` with `x y z` all being arguments for --show
instead of `x` only.

Maybe there are workarounds with other flag-parsing libraries. Something
to look at in the future.

Fixes: #83
Signed-off-by: Valentin Rothberg <rothberg@redhat.com>