grep output: force printing filename for grep/ripgrep
In the case where a single file to search is passed to ripgrep (for
example: 'vgrep my_pattern foo.c'), ripgrep does not print the name of
the file in its output. Then vgrep tries to split the result string on
the NUL byte and to access the second substring, but panics because it
does not exist:

    panic: runtime error: index out of range [1] with length 1

This case never happened until recently because we would happen the
current working directory ('.') to the list of files to search, but this
was changed in a recent commit.

Fix by forcing grep and ripgrep to print the name of the file, even if
there is just one file to search.

Add a series of test to make sure we can grep with the different tools
by passing a directory or a file name.

Fixes: #103
Fixes: 3738a019c0ec ("grep commands: Do not pass '.' as an argument to grep/ripgrep")