grep output: expect no more than 2 substrings when splitting on NUL-byte
When splitting grep and ripgrep output on the NUL-byte separator, we
expect up to three substrings (the third one would contain the end of
the original string, regardless of additional, unused NUL bytes).

But for grep/ripgrep we really want only two substrings, one for the
file path, and one for the line number + colon separator + match result.

While this has no incidence in most cases, it might affect the output if
the user passes the '-a' option to grep or ripgrep. Let's change the
number of expected substrings.

Here is an example with a file containing NUL bytes between the square
brackets. Before:

    $ cat foo/test
    FOOBAR [] NUL-bytes between brackets [] and once more []

    $ vgrep -a FOOBAR foo/
    Index File     Line Content
        0 foo/test    1 FOOBAR [

After:

    $ vgrep -a FOOBAR foo/
    Index File     Line Content
        0 foo/test    1 FOOBAR [^@] NUL-bytes between brackets [^@] and once more [^@]

Do not change the number of expected substrings for git grep, as 3 is
the value we want in that case.

Add a series of test to make sure we get the end of lines with
NUL-bytes.