git grep: Run 'git rev-parse' only if not selecting ripgrep
Vgrep always runs 'git rev-parse' to check whether the current directory
is a git repository. This is useless if ripgrep is installed, since it
is picked by default. Let's run git only if ripgrep is not selected.

Add some tests to check that all expected grep tools are picked,
depending on the circumstances. They rely on the debug output to inspect
what command is run.