delete: fix index update for multi-line deletions
The operation used to update the index when deleting a selection of
lines is incorrect. Deletions are done sequentially, so after the first
deletion, the indices in the list of results have changed and we need to
take that into account. But each time we remove an entry, the list of
results _shrinks_ by one, so we want a _smaller_ index. The addition in
"index := idx + offset" is wrong, it should be a subtraction. Let's fix.

Add a test to validate the results, and document the missing "delete"
command in the README.md.