interactive: switch from bufio Scanner to (readline-like) liner
Use the readline Go library to process commands in interactive mode. It
provides better support for input (such as directional arrows) and a
number of classic readline shortcuts [0]. It also does command history,
which can be handy for editing the range of a previous command for
example. Autocompletion also exists but is not implemented in this
commit (the one-letter commands hardly need any completion).

Note that we loose bold marking on the prompt (or the ability to use
other ANSI control codes for the prompt in the future), because the
library explicitly disable it to avoid compatibility issues with Windows
[1].

[0] https://github.com/peterh/liner#line-editing
[1] https://github.com/peterh/liner/commit/160e3795572865ddd53357b3c3744fef57925885