colwriter: prevent panic on --no-header and no result found
If we pass the --no-header option _and_ if no results are to be
displayed (because no match was found or because we filtered them out
in the interactive shell), the ColWriter attempts to use rows[0] which
does not exist, thus causing vgrep to panic. Let's return early if
len(rows) is nul instead.