interactive: close command-line editor on early os.Exit()
If we fail to get input from the prompt, typically because the user
types Ctrl-C or sends EOF (e.g. Ctrl-D), then the defer to close the
readline-like command line editor is not executed, leaving the terminal
in a Weird State. I'm not sure what is changed exactly, but later
calling other command-line prompts (e.g. "git add -p", typing a SSH
password) will have the "Enter" key not behaving correctly.

Fix by closing the object before exiting. Looking at the code in
vgrep.go, it seems that all other calls to os.Exit() are in the main
function (i.e. not between liner.NewLiner() and return from
commandParse()), so we should not hit this issue elsewhere. It may still
happen if vgrep panics (but then that would likely be another bug in
itself). In such a case, note that typing "reset" in the terminal resets
the state and also fix the issue.

Fixes: #106
Fixes: 88894b1c7b7e ("interactive: switch from bufio Scanner to (readline-like) liner")