(new) grep: add a command to trigger a new search from interactive shell
A possible workflow with vgrep is as follows:

  1. Vgrep from terminal command line
  2. Enter the interactive shell
  3. Filter entries, print context, show entries, etc.
  4. Exit the interactive shell
  5. Launch a new vgrep search from terminal command line
  6. Enter the interactive shell again
  7. Work on the new set of results, etc.

Starting a new search in step 5 requires to exit then enter again the
interactive shell (steps 4 and 6). We could have a shortcut to prevent
that and to trigger a new search from within the interactive shell.

This commit does just that. The new command "g"/"grep" for the shell
takes a set of arguments and pass them to a new vgrep search, without
leaving the shell. The results for the new search are printed
immediately.

Arguments are parsed as shell words (simply parsing on spaces would not
process arguments enclosed in quotes in a correct fashion).

Vgrep's cache is updated with the results from the new query.

Related test is included.