(new) grep: add filename completion
Completion for files and directory names is great to have when typing
grep commands [citation needed]. This commit adds autocompletion for
filenames for the "grep" command in the interactive shell.

- Since we add completion, implement completion for vgrep command names.
- Other than command names, completion is only provided for the "grep"
  command.
- Completion for "grep" only starts for the third argument (first is
  command name, second one is match pattern).
- It does not look perfect yet but that's probably a decent start.

Move the list of command used in the help message to a reusable
variable, to avoid to duplicate it.

I could not find an easy way to use completion in the test framework, so
no test at this point.