splitMatch: fail gracefully on failure to parse output
Vgrep may fail to parse the output from the selected grep tool, in
particular if the user provided some options that modify the output
(-A, -B, -C, -a, ...). In such a case, splitting the results on the
separators will not produce the expected number of substrings and will
result in attempts to access incorrect indices in the slice.

Fix splitMatch() so vgrep exits gracefully with a hint message when it
cannot parse the output.

Add a test for this case.