selectors: add a "all" selector to process all results
Some vgrep commands ("print", "context" for example) defaults to
processing all search results if no selector is passed. Some other
commands ("show", "delete") are slightly more sensitive, and they
require an explicit set of selectors to indicate the entries to process.

For those commands, it could be interesting to have a shortcut to tell
them "do process all entries, I know what I am doing". This commit
implements a new selector, "all", that does just that.

A new test file is added to check that selectors are parsed as expected.