main: exit with 1 if there are no matches
This mimics the behavior of grep/git-grep and ripgrep, and is useful
for chaining a search with a later command:

    vgrep "matches zero lines" && vgrep --interactive

As an implementation detail, this commit also reworks the final stages
of main(), for two reasons:

- to more easily handle in which cases it is necessary to Wait();
- to fix a probable (though not confirmed)[1] bug where a search that
resulted in 0 matches would cause the behavior of later commands to be
undefined, due to cacheWrite not necessarily completing before the
program terminated.

[1] In a simple test cacheWrite does not complete and leaves the
previous search results in place.